Removed Plane Extract Ransac from tilting lidar and made minor modifications to launch and cpp files

+++ tilting_lidar_clean/hokuyo_ground_filter.cpp
@@ -39,7 +39,7 @@ void filterGround (const sensor_msgs::PointCloud2ConstPtr& msg)
  
    int n = cloud.width;
 
-   int tolerance = max_z + 0.1;
+   float tolerance = max_z + 0.1;
 
     aux.points.resize(n);
